Remove (int) casting for a phone type enum.
The phonetypes saved in the DB are single letters so we could return these instead rather than an integer value.

+++ UHResidentInformationAPI.Tests/V1/Gateways/UHGatewayTests.cs
@@ -89,7 +89,7 @@ namespace UHResidentInformationAPI.Tests.V1.Gateways
 
             var databasePhoneEntity = TestHelper.CreateDatabaseTelephoneNumberForPersonId(databaseContactLink.ContactID);
 
-            var type = (int) PhoneType.X;
+            var type = PhoneType.X;
             databasePhoneEntity.Type = type.ToString();
 
             UHContext.TelephoneNumbers.Add(databasePhoneEntity);
